Backend/requirements-test.txt
pytest==7.4.0
requests==2.31.0
bcrypt==4.0.1
jsonschema==4.20.0
pytest-cov==4.1.0
python-dotenv==1.0.0